change setFloat to setDouble on android to get right values

+++ CBLClient/Apps/CBLTestServer-Android/app/src/main/java/com/couchbase/CouchbaseLiteServ/server/ValueSerializer.java
@@ -100,7 +100,7 @@ public class ValueSerializer {
         } else if (value.startsWith("L")){
             return (T) new Long(value.substring(1));
         } else if (value.startsWith("F")){
-            return (T) new Float(value.substring(1));
+            return (T) new Double(value.substring(1));
         } else if (value.startsWith("D")){
             return (T) new Double(value.substring(1));
         }
